add --dry-run, -o might be working

+++ function_runProgram.py
@@ -8,7 +8,7 @@ from function_getMetadata import getAndSaveMetadata
 from function_addMetadataAndMaps import addMetadataAndMaps
 
 
-def runProgram(filename, outputFileName, filenameAndDirectory, iterations, failedFiles, warningFiles, infoMessages, outputFileNameAndDirectory, currentOS, engAudioNoSubs, forceOverwrite, skippedFiles):
+def runProgram(filename, outputFileName, filenameAndDirectory, iterations, failedFiles, warningFiles, infoMessages, outputFileNameAndDirectory, currentOS, engAudioNoSubs, forceOverwrite, skippedFiles, dryRun):
 
     # Check If File Exists
     if forceOverwrite is False:
@@ -33,40 +33,44 @@ def runProgram(filename, outputFileName, filenameAndDirectory, iterations, faile
     # Process metadata
     metadataAndMaps, infoMessages = addMetadataAndMaps(filenameAndDirectory, metadataTable, totalNumOfStreams, currentOS, engAudioNoSubs, infoMessages)
 
-    print(Fore.CYAN + "Started: " + filename + Fore.RESET, end="\r")  # Print and return courser to the start of the line
+    if dryRun is False:
+        print(Fore.CYAN + "Started: " + filename + Fore.RESET, end="\r")  # Print and return courser to the start of the line
 
-    # Time for FFmpeg to do its thing:
-    if currentOS == "Linux":
-        #print("ffmpeg " + overwriteOption + " -v error -xerror -i \"" + filenameAndDirectory + "\" -map_metadata -1 -map_chapters 0" +  metadataAndMaps + " -metadata title=\"\" -c copy \"" + outputFileNameAndDirectory + "\"")
-        errorCheck = run("ffmpeg " + overwriteOption + " -v error -xerror -i \"" + filenameAndDirectory + "\" -map_metadata -1 -map_chapters 0" + metadataAndMaps + " -metadata title=\"\" -c copy \"" + outputFileNameAndDirectory + "\"", capture_output=True, shell=True)
-    elif currentOS == "Windows":
-        #print("ffmpeg " + overwriteOption + " -v error -xerror -i \"" + filenameAndDirectory + "\" -map_metadata -1 -map_chapters 0" + metadataAndMaps + " -metadata title=\"\" -c copy \"" + outputFileNameAndDirectory + "\"")
-        errorCheck = run("cmd /c ffmpeg " + overwriteOption + " -v error -xerror -i \"" + filenameAndDirectory + "\" -map_metadata -1 -map_chapters 0" + metadataAndMaps + " -metadata title=\"\" -c copy \"" + outputFileNameAndDirectory + "\"", capture_output=True, shell=True)
+        # Time for FFmpeg to do its thing:
+        if currentOS == "Linux":
+            #print("ffmpeg " + overwriteOption + " -v error -xerror -i \"" + filenameAndDirectory + "\" -map_metadata -1 -map_chapters 0" +  metadataAndMaps + " -metadata title=\"\" -c copy \"" + outputFileNameAndDirectory + "\"")
+            errorCheck = run("ffmpeg " + overwriteOption + " -v error -xerror -i \"" + filenameAndDirectory + "\" -map_metadata -1 -map_chapters 0" + metadataAndMaps + " -metadata title=\"\" -c copy \"" + outputFileNameAndDirectory + "\"", capture_output=True, shell=True)
+        elif currentOS == "Windows":
+            #print("ffmpeg " + overwriteOption + " -v error -xerror -i \"" + filenameAndDirectory + "\" -map_metadata -1 -map_chapters 0" + metadataAndMaps + " -metadata title=\"\" -c copy \"" + outputFileNameAndDirectory + "\"")
+            errorCheck = run("cmd /c ffmpeg " + overwriteOption + " -v error -xerror -i \"" + filenameAndDirectory + "\" -map_metadata -1 -map_chapters 0" + metadataAndMaps + " -metadata title=\"\" -c copy \"" + outputFileNameAndDirectory + "\"", capture_output=True, shell=True)
 
-    if len(str(errorCheck.stderr)) > 8:  # Integrity and error check
-        if str(errorCheck.stderr).find("Referenced QT chapter track not found") != -1:
-            print(filename + Fore.CYAN + ": Was not encoded to specification" + Fore.RESET)
-            infoMessages += 1
-        elif str(errorCheck.stderr).find("already exists. Exiting.") != -1:
-            print("already exists")
-        else:
-            errorOutput = re.sub("b\"", "", str(errorCheck.stderr))
-            errorOutput = re.sub("b\'", "", errorOutput)
-            errorOutput = re.sub(r"\\r\\n\'", "", errorOutput)
-            errorOutput = re.sub(r"\\r\\n\"", "", errorOutput)
-            errorOutput = re.sub(r"\\r\\n", "", errorOutput)
-            print(Fore.YELLOW + "FFmpeg Error: " + "\"" + Fore.RESET + errorOutput + Fore.YELLOW + "\"" + Fore.RESET)
+        if len(str(errorCheck.stderr)) > 8:  # Integrity and error check
+            if str(errorCheck.stderr).find("Referenced QT chapter track not found") != -1:
+                print(filename + Fore.CYAN + ": Was not encoded to specification" + Fore.RESET)
+                infoMessages += 1
+            elif str(errorCheck.stderr).find("already exists. Exiting.") != -1:
+                print("already exists")
+            else:
+                errorOutput = re.sub("b\"", "", str(errorCheck.stderr))
+                errorOutput = re.sub("b\'", "", errorOutput)
+                errorOutput = re.sub(r"\\r\\n\'", "", errorOutput)
+                errorOutput = re.sub(r"\\r\\n\"", "", errorOutput)
+                errorOutput = re.sub(r"\\r\\n", "", errorOutput)
+                print(Fore.YELLOW + "FFmpeg Error: " + "\"" + Fore.RESET + errorOutput + Fore.YELLOW + "\"" + Fore.RESET)
 
-            packingSpaces = ' ' * (len(Fore.BLUE + "Started: " + filename + Fore.RESET) - len(Fore.YELLOW + "May Be Corrupted: " + filename + Fore.RESET))  # Pack the output with spaces or there will be characters left from the overwritten print
-            print(Fore.YELLOW + "May Be Corrupted: " + Fore.RESET + filename + Fore.YELLOW + ", so it was not copied" + Fore.RESET + packingSpaces)
-            warningFiles += 1
-            iterations -= 1  # Remove from successful count
-            try:
-                os.remove(Path(outputFileNameAndDirectory))  # Delete the file since its corrupted
-            except:
-                print(":o Could not delete " + outputFileNameAndDirectory)
-            return iterations, failedFiles, warningFiles, infoMessages, skippedFiles
+                packingSpaces = ' ' * (len(Fore.BLUE + "Started: " + filename + Fore.RESET) - len(Fore.YELLOW + "May Be Corrupted: " + filename + Fore.RESET))  # Pack the output with spaces or there will be characters left from the overwritten print
+                print(Fore.YELLOW + "May Be Corrupted: " + Fore.RESET + filename + Fore.YELLOW + ", so it was not copied" + Fore.RESET + packingSpaces)
+                warningFiles += 1
+                iterations -= 1  # Remove from successful count
+                try:
+                    os.remove(Path(outputFileNameAndDirectory))  # Delete the file since its corrupted
+                except:
+                    print(":o Could not delete " + outputFileNameAndDirectory)
+                return iterations, failedFiles, warningFiles, infoMessages, skippedFiles
 
-    packingSpaces = " " * (len(Fore.BLUE + "Started: " + filename + Fore.RESET) - len(Fore.GREEN + "Done: " + outputFileName + Fore.RESET))  # Pack the output with spaces or there will be characters left from the overwritten print
-    print(Fore.GREEN + "Done: " + outputFileName + Fore.RESET + packingSpaces)
+        packingSpaces = " " * (len(Fore.BLUE + "Started: " + filename + Fore.RESET) - len(Fore.GREEN + "Done: " + outputFileName + Fore.RESET))  # Pack the output with spaces or there will be characters left from the overwritten print
+        print(Fore.GREEN + "Done: " + outputFileName + Fore.RESET + packingSpaces)
+    else:
+        print("Example change: " + filename + " --> " + outputFileName)
+
     return iterations, failedFiles, warningFiles, infoMessages, skippedFiles
